# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Recursos
# during dev, should point to your local machine
host: localhost:10010

# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /resource:
    # binds a127 app logic to a route
    x-swagger-router-controller: resourcetype.controller
    get:
      description: Obtiene la lista de recursos
      # used as the method name of the controller
      operationId: getResourceType
      
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetResourceTypeListResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post: 
      description: Agregar un tipo a la lista
      operationId: createResourceType
      parameters: 
        - in: body
          name: resourcetype
          description: El recurso a crear
          schema: 
            $ref: "#/definitions/ResourceType"
      responses: 
        "201": 
          description: Success
          schema: 
            $ref: "#/definitions/GetResourceTypeResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:

  ResourceType:
    type: object
    properties: 
      name: 
        type: string
        description: Nombre del tipo de recurso
    required: 
      - name
  GetResourceTypeResponse:
    type: object
    properties: 
      id:
        type: number
        description: identifier
      name:
        type: string
        description: Nombre del tipo de recurso
    required: 
      - name
  GetResourceTypeListResponse:
    required:
      - resourcetype
    properties:
      resourcetype:
        type: array
        items:
          $ref: "#/definitions/GetResourceTypeResponse"
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string